# swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
    "AudioMetaData": {
      "properties": {
        "durationInMillis": {
          "required": false,
          "type": "integer"
        },
        "normalizedLoudness": {
          "items": {
            "type": "integer"
          },
          "required": false,
          "type": "array"
        }
      },
      "type": "object"
    },
...
    "ErrorMessage": {
      "properties": {
        "code": {
          "type": "integer"
        },
        "error": {
          "type": "string"
        }
      },
      "required": ["code", "error"],
      "type": "object"
    },
    "Image": {
      "properties": {
        "data": {
          "format": "base64",
          "type": "string"
        },
        "expectsReadConfirmation": {
          "type": "boolean"
        },
        "height": {
          "type": "integer"
        },
        "legalHoldStatus": {
          "$ref": "#/definitions/LegalHoldStatus"
        },
        "messageTimer": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "width": {
          "type": "integer"
        }
      },
      "required": ["conversationId", "data", "height", "type", "width"],
      "type": "object"
    },
    "Instance": {
      "allOf": [
        {
          "$ref": "#/definitions/InstanceAndName"
        },
        {
          "properties": {
            "backend": {
              "type": "string"
            },
            "clientId": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
    "InstanceAndName": {
      "properties": {
        "instanceId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "LegalHoldStatus": {
      "description": "Type can be `0` (unknown), `1` (disabled) or `2` (enabled).",
      "enum": [0, 1, 2],
      "type": "integer"
    },
...
    "ServerErrorMessage": {
      "allOf": [
        {
          "$ref": "#/definitions/ErrorMessage"
        },
        {
          "properties": {
            "stack": {
              "type": "string"
            }
          },
          "type": "object"
        }
      ]
    },
...
    "VideoMetaData": {
      "properties": {
        "durationInMillis": {
          "required": false,
          "type": "integer"
        },
        "height": {
          "required": false,
          "type": "integer"
        },
        "width": {
          "required": false,
          "type": "integer"
        }
      },
      "type": "object"
    }
  },
  "info": {
    "contact": {
      "email": "opensource@wire.com"
    },
    "description": "End-to-end Test Service (ETS) for Wire's test automation suite.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/wireapp/wire-web-ets/blob/master/LICENSE"
    },
    "termsOfService": "https://wire.com/legal/",
    "title": "E2E Test Service",
    "version": "1.20.0"
  },
  "paths": {
...
    "/instance/{instanceId}": {
      "delete": {
        "description": "",
        "operationId": "deleteInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Delete an instance",
        "tags": ["Instance"]
      },
      "get": {
        "operationId": "getInstance",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/Instance"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get information about an instance",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/fingerprint": {
      "get": {
        "operationId": "getFingerprint",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "properties": {
                "fingerprint": {
                  "type": "string"
                },
                "instanceId": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Get the fingerprint from the instance's client",
        "tags": ["Instance"]
      }
    },
...
    "/instance/{instanceId}/sendFile": {
      "post": {
        "operationId": "sendFile",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "audio": {
                  "$ref": "#/definitions/AudioMetaData"
                },
                "conversationId": {
                  "format": "uuid",
                  "type": "string"
                },
                "data": {
                  "format": "base64",
                  "type": "string"
                },
                "expectsReadConfirmation": {
                  "type": "boolean"
                },
                "fileName": {
                  "type": "string"
                },
                "legalHoldStatus": {
                  "$ref": "#/definitions/LegalHoldStatus"
                },
                "messageTimer": {
                  "type": "integer"
                },
                "type": {
                  "type": "string"
                },
                "video": {
                  "$ref": "#/definitions/VideoMetaData"
                }
              },
              "required": ["conversationId", "data", "fileName", "type"],
              "type": "object"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/InstanceAndName"
                },
                {
                  "properties": {
                    "messageId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  }
                }
              ]
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Send a file to a conversation",
        "tags": ["Instance"]
      }
    },
    "/instance/{instanceId}/sendImage": {
      "post": {
        "operationId": "sendImage",
        "parameters": [
          {
            "description": "ID of instance to use",
            "format": "uuid",
            "in": "path",
            "name": "instanceId",
            "required": true,
            "type": "string"
          },
          {
            "description": "",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "allOf": [
                {
                  "properties": {
                    "conversationId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": ["conversationId"]
                },
                {
                  "$ref": "#/definitions/Image"
                }
              ]
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/InstanceAndName"
                },
                {
                  "properties": {
                    "messageId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  }
                }
              ]
            }
          },
          "404": {
            "description": "Instance not found",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ServerErrorMessage"
            }
          }
        },
        "summary": "Send an image to a conversation",
        "tags": ["Instance"]
      }
    },
...
  },
  "schemes": ["http"],
  "swagger": "2.0",
...
}